+++ src/api/domain/user-access/user/user-access.controller.ts
@@ -3,10 +3,12 @@ import { Router, RequestHandler } from 'express';
 import { loginActionValidation } from './actions/login.action';
 import { authorizationMiddleware } from '@api/middlewares/auth/auth.middleware';
 import { updateUserPasswordValidation } from './actions/update-user-password.action';
+import { refreshTokenActionValidation } from './actions/refresh-token.action';
 
 interface Dependencies {
   updateUserPasswordAction: RequestHandler;
   loginAction: RequestHandler;
+  refreshTokenAction: RequestHandler;
 }
 
 export class UserAccessController extends Controller {
@@ -22,6 +24,12 @@ export class UserAccessController extends Controller {
       '/update-password',
       [authorizationMiddleware, updateUserPasswordValidation],
       this.dependencies.updateUserPasswordAction,
+    );
+
+    router.post(
+      '/refresh-token',
+      [refreshTokenActionValidation],
+      this.dependencies.refreshTokenAction,
     );
 
     return router;